uart-python/requirements.txt
pyserial==3.4
pytest==3.9.3